requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0
Pillow==10.4.0
